This commit include the following new features
- Support for fw_printenv and fw_setenv
- Support for OF.LOCO101 opeframe board
- New linux-kernel programming script
- New u-boot programming script with autosave environment
- U-boot MAC address fixed bug in LTIB mode
- Fixed maximum frequency for industrial modules to 800Mhz
- WF111 recipies improvment
- G150 LCD driver improvment

+++ recipes-core/engicam-emmc-script/engicam-emmc-script/q7_tftp_boot.sh
@@ -1,13 +1,20 @@
-serverip=192.168.2.96
-filename=u-boot.imx
-tftp -g -r $filename -l $filename $serverip
-if [ -f $filename ]; then
- 	echo "Download " $filename
-	#dd if=$filename of=u-boot_my.imx bs=512 seek=2 
-	dd if=$filename of=/dev/mmcblk1 bs=512 seek=2
-	#rm u-boot_my.imx
-	rm $filename
-	sync
+if env | grep -q ^serverip=; then
+	filename=u-boot.imx
+	echo "Download " $filename
+	tftp -g -r $filename -l $filename $serverip
+	if [ -f $filename ]; then
+		fw_printenv > parametri.txt
+		cat parametri.txt | sed -e 's/=/ /' > convertito.txt
+		dd if=$filename of=/dev/mmcblk1 bs=512 seek=2
+		rm $filename
+		fw_setenv -s convertito.txt
+		rm parametri.txt
+		rm convertito.txt
+		sync
+	else
+		echo "Unable to download " $filename
+	fi
 else
-	echo "Unable to download " $filename
+	echo "Serverip not setted:"
+	echo "    export serverip=192.168.XXX.XXX"
 fi
